requirements.txt
# Core
boto3>=1.28.0
click>=8.1.0
pyyaml>=6.0.0
python-dateutil>=2.8.0

# Database
duckdb>=0.9.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Export
jinja2>=3.1.0

# Security
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0
